refactor(*): use PascalCase in component
