+++ app/features/Pairing/Portal/index.tsx
@@ -1,5 +1,5 @@
 import { usePairContext } from '../context';
-import { HeaderPortal } from '~/components/layout/header';
+import { HeaderPortal } from '~/components/Layout/header';
 import IconButton from '~/components/IconButton';
 import { Fragment } from 'react';
